Rename to `signAndSend`
